Update libfwsi to version 20220927 / rev 15 via SR 1038478
https://build.opensuse.org/request/show/1038478
by user jengelh + dimstar_suse
delete system-libs.patch

+++ packages/libf/libfwsi/libfwsi.spec
@@ -28,11 +28,8 @@ Source:         https://github.com/libyal/libfwsi/releases/download/%version/lib
 Source2:        https://github.com/libyal/libfwsi/releases/download/%version/libfwsi-experimental-%version.tar.gz.asc
 Source3:        %name.keyring
 Source9:        Windows_Shell_Item_format.pdf
-Patch1:         system-libs.patch
 BuildRequires:  %{python_module devel}
 BuildRequires:  c_compiler
-BuildRequires:  gettext-tools >= 0.21
-BuildRequires:  libtool
 BuildRequires:  pkg-config
 BuildRequires:  python-rpm-macros
 BuildRequires:  pkgconfig(libcdata) >= 20220115
@@ -46,6 +43,7 @@ BuildRequires:  pkgconfig(libfole) >= 20220115
 BuildRequires:  pkgconfig(libfwps) >= 20220122
 BuildRequires:  pkgconfig(libuna) >= 20220611
 %python_subpackages
+# Various notes: https://en.opensuse.org/libyal
 
 %description
 Library to access the Windows Shell Item format for the libyal family of libraries.
@@ -77,22 +75,19 @@ applications that want to make use of libfwsi.
 cp %_sourcedir/*.pdf .
 
 %build
-autoreconf -fi
-# OOT builds are presently broken, so we have to install
-# within each python iteration now, not in %%install.
 %{python_expand #
-# see libcdata for version-sc
 echo "V_%version { global: *; };" >v.sym
 %configure --disable-static \
 	--enable-python PYTHON_VERSION="%{$python_bin_suffix}" \
 	LDFLAGS="-Wl,--version-script=$PWD/v.sym"
+grep ' '' ''local' config.log && exit 1
 %make_build
 %make_install DESTDIR="%_builddir/rt"
 %make_build clean
 }
 
 %install
-mv %_builddir/rt/* %buildroot/
+mv "%_builddir/rt"/* %buildroot/
 find %{buildroot} -type f -name "*.la" -delete -print
 
 %post   -n %{lname} -p /sbin/ldconfig
